Handle setting $Input and $InputFileName during Get[] (#1011)
Currently the global variables `$Input` and `$InputFileName` are not set
correctly. This pull request fixes this.

### Description of the bug
#### Steps to reproduce
- Create a file `script.m` with contents  
```
Print["Input: ",$Input]
Print["InputFileName: ", $InputFileName]
```
- Within mathics execute: `Get["script.m"]`
- or call the script from the command line `mathics script.m`

#### Observed output on mathics as of commit
430b1f47840fa064a8e324f8f2b759fb1db1bef3
Using `Get` the global variables are not set at all.
```
In[1]:= Get["script.m"]
Script Input: 
Script InputFileName: 
Out[1]= None
```

From the command line `InputFileName` is set, but not the full path.
`Input` is still unset.
```
Script Input: 
Script InputFileName: script.m
```

#### Expected output
```
In[1]:= Get["script.m"]                                                        
Input: script.m
InputFileName: /full/path/to/script.m
```
