+++ mathics/builtin/files_io/files.py
@@ -13,6 +13,7 @@ from io import BytesIO
 from mathics_scanner import TranslateError
 
 import mathics
+import mathics.eval.files_io.files as eval_files
 from mathics.core.atoms import Integer, String, SymbolString
 from mathics.core.attributes import A_PROTECTED, A_READ_PROTECTED
 from mathics.core.builtin import (
@@ -53,12 +54,16 @@ INPUT_VAR = ""
 
 SymbolInputStream = Symbol("InputStream")
 SymbolOutputStream = Symbol("OutputStream")
-SymbolPath = Symbol("$Path")
 
 # TODO: Improve docs for these Read[] arguments.
 
 # ## FIXME: All of this is related to Read[]
 # ## it can be moved somewhere else.
+
+
+def set_input_var(input_string):
+    global INPUT_VAR
+    INPUT_VAR = input_string
 
 
 class Input_(Predefined):
@@ -387,54 +392,25 @@ class Get(PrefixOperator):
     def eval(self, path, evaluation: Evaluation, options: dict):
         "Get[path_String, OptionsPattern[Get]]"
 
-        def check_options(options):
-            # Options
-            # TODO Proper error messages
-
-            result = {}
-            trace_get = evaluation.parse("Settings`$TraceGet")
-            if (
-                options["System`Trace"].to_python()
-                or trace_get.evaluate(evaluation) is SymbolTrue
-            ):
-                import builtins
-
-                result["TraceFn"] = builtins.print
-            else:
-                result["TraceFn"] = None
-
-            return result
-
-        py_options = check_options(options)
-        trace_fn = py_options["TraceFn"]
-        result = None
-        pypath = path.get_string_value()
+        # wrap actual evaluation to handle setting $Input
+        # and $InputFileName
+        global INPUT_VAR
         definitions = evaluation.definitions
-        mathics.core.streams.PATH_VAR = SymbolPath.evaluate(evaluation).to_python(
-            string_quotes=False
-        )
+        # store input paths of calling context
+        outer_input_var = INPUT_VAR
+        outer_inputfile = definitions.get_inputfile()
+        # set new input paths
+        pypath = path.get_string_value()
+        INPUT_VAR = pypath
+        definitions.set_inputfile(pypath)
+
+        # perform the actual evaluation
         try:
-            if trace_fn:
-                trace_fn(pypath)
-            with MathicsOpen(pypath, "r") as f:
-                feeder = MathicsFileLineFeeder(f, trace_fn)
-                while not feeder.empty():
-                    try:
-                        query = parse(definitions, feeder)
-                    except TranslateError:
-                        return SymbolNull
-                    finally:
-                        feeder.send_messages(evaluation)
-                    if query is None:  # blank line / comment
-                        continue
-                    result = query.evaluate(evaluation)
-        except IOError:
-            evaluation.message("General", "noopen", path)
-            return SymbolFailed
-        except MessageException as e:
-            e.message(evaluation)
-            return SymbolFailed
-        return result
+            return eval_files.eval_Get_inner(self, path, evaluation, options)
+        finally:
+            # always restore input paths of calling context
+            INPUT_VAR = outer_input_var
+            definitions.set_inputfile(outer_inputfile)
 
     def eval_default(self, filename, evaluation):
         "Get[filename_]"
